+++ packages/cli-plugin-router/template/template-react-ts/src/router/index.tsx
@@ -1,6 +1,7 @@
 import React, { lazy } from 'react';
 import { Navigate, RouteObject } from 'react-router-dom';
 
-const routes: RouteObject[] = [];
+const routes: RouteObject[] = [
+];
 
 export default routes;