+++ packages/shared-utils/lib/module.ts
@@ -2,10 +2,10 @@ import path from 'path';
 
 export default async (request: string, targetDir: string) => {
   const res = await import(
-    path.resolve(
+    `file:///${path.resolve(
       targetDir,
-      `./node_modules/@micro-cli/${request.slice(7)}/dist/index.mjs`
-    )
+      `./node_modules/@micro-cli/${request.slice(11)}/dist/index.mjs`
+    )}`
   );
   return res;
 };