+++ packages/add/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @micro-cli/add
 
+## 1.0.5
+
+### Patch Changes
+
+- publish test
+- Updated dependencies
+  - @micro-cli/shared-utils@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes