+++ packages/cli-plugin-typescript/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @micro-cli/cli-plugin-typescript
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @micro-cli/create@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- publish test
+- Updated dependencies
+  - @micro-cli/create@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes