+++ packages/core/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @micro-cli/core
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @micro-cli/create@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- publish test
+- Updated dependencies
+  - @micro-cli/add@1.0.5
+  - @micro-cli/create@1.0.5
+  - @micro-cli/shared-utils@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes