+++ packages/create/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @micro-cli/create
 
+## 1.0.6
+
+### Patch Changes
+
+- change download path
+  - @micro-cli/cli-plugin-eslint@1.0.6
+  - @micro-cli/core@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- publish test
+- Updated dependencies
+  - @micro-cli/cli-plugin-eslint@1.0.5
+  - @micro-cli/core@1.0.5
+  - @micro-cli/shared-utils@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes